migrate MMRVerification test
